♻️  refactor: move fullscreen external store to lib/fullscreen (#27)
